test-mesa: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@42641 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ test/mesa/mesasimplerendering.c
@@ -408,7 +408,7 @@ int main(void)
         {
             /* FPS counting is naive! */
             fpsmicrosecs += 1000000;
-            sprintf(title, "MesaSimpleRendering, FPS: %d", fps);
+            sprintf(title, "MesaSimpleRendering, FPS: %d", (int)fps);
             SetWindowTitles(win, title, (UBYTE *)~0L);
             fps = 0;
         }
